Add admin authorization to delete ticket

+++ src/main/java/com/ercanbeyen/bloggingplatform/service/impl/TicketServiceImpl.java
@@ -73,6 +73,7 @@ public class TicketServiceImpl implements TicketService {
 
     @Override
     public String deleteTicket(Integer id) {
+        SecurityUtil.checkAdminRole();
         ticketMapper.deleteTicketById(id);
         return String.format(ResponseMessage.SUCCESS, EntityName.TICKET, id, ResponseMessage.Operation.DELETED);
     }
